The actual replacement of the AndroidApiWrapper.
Sorry, forgot to add this to my previous commit and already pushed.

Anyway, here the single function that was used by our code that was implemented
in the AndroidApiWrapper has been replaced with the equivalent SDL call (which
is Android specific). As we cannot run this code yet, it is not tested, only
that the openmsx.so can be successfully built now.