Optimize adding multiple elements to a Tcl list
This is a follow-up on d3a23f2abf13a8.

Before this patch we added elements to a Tcl list one-by-one. That is for each
element we made a call into the libtcl library. Now we add multiple elements at
once. That is with a single call to libtcl.

This is similar to adding elements to a vector one-by-one (repeated calls to
vector::push_back()) or multiple elements at once (e.g. with a call to a range
vector::insert() overload).

And the advantage for Tcl lists is similar than the advantage for std::vectors,
the internal memory buffer only has to be reallocated at most once.