[2/7] Don't invalidate CPUCore cache-lines on slot switch
The previous commit message described why we need to invalidate cache lines
whenever the Z80 address space layout changes. A typical example of when this
happens on MSX is when the primary or secondary slot select registers are
written. It turns out this happens relatively often. For example in MSX-DOS
when an IRQ occurs, we switch page 0 from RAM to BIOS-ROM and back. So 100 or
more switches per second is not uncommon.

This patch tries to mitigate the cost of slot switching. To do this we keep 16
(=4*4) tables with cache line information, one for each primary/second slot
combination. When we switch slot, we also switch to the correct set of
cache-lines.

The obvious way to implement this is to put those 16 tables in CPUCore and on
the memory read/write paths pick the correct table depending on the selected
slot. Though that would change the fast memory-access-path from
    const byte* line = readCacheLine[address >> 8];
    if (line) return line[address]
into
    const byte* line = readCacheLine[selectedSlot[address >> 14]][address >> 8];
    if (line) return line[address]
But we don't want any extra code on this path. So instead we do the following.

CPUCore still doesn't know anything about slots, so it continues using the same
(single set) of cache-lines as before. But the MSXCPU class has 16 additional
shadow cache-line sets (thus there are 16+1=17 sets in total). (Only) the set
in CPUCore is kept up-to-date (more specifically: only these can switch from
'unknown' to either 'uncacheable' or 'some backdoor pointer'). When we switch
slot, we first copy (part of) the CPUCore set to the shadow set in MSXCPU, and
then copy (part of) another shadow set back to CPUCore.

The only tricky part are the remaining cache-line invalidations (there are still
other reasons than slot switching to invalidate). These now also have to
invalidate the shadow sets. In later patches I hope to also mitigate (the most
common of) those cache-line invalidations (e.g. memory mapper or megarom segment
switches).

This patch resulted in a 3.5% speedup in the zexall benchmark (no video/sound
rendering, mostly only Z80 emulation): my test went from 34.44 -> 33.23 seconds
realtime to emulate 16700 seconds MSX time. First time I achieve more than 500x
realtime speed!

Here are the updated profile counters (compare to the results in the first
patch of this series):

Zexall
  NonCachedRead: 691483
  NonCachedWrite: 154259
  GetReadCacheLine: 355
  GetWriteCacheLine: 212
  SlowRead: 691154
  SlowWrite: 154067
  DisallowCacheRead: 52698
  DisallowCacheWrite: 104479
  InvalidateCache: 74
 This is a MSX-DOS application which does slot-switching 2x50 times per second
 in the ISR (switch from RAM to BIOS-ROM and back). Most 'InvalidateCache'
 events have disappeared. Also there's a great reduction of
 'NonCachedRead/Write' events, and an even more drastic reduction for
 'GetRWCacheLine'. In fact the numer of remaining 'NonCachedRW' events is
 almost equal to the number of 'SlowRW' and 'DisallowCacheRW' events, this
 means the cache is working close to optimal (hence the 3% speedup reported
 above).

Pumpkin adventure 3, intro demo
  NonCachedRead: 1971753
  NonCachedWrite: 180320
  GetReadCacheLine: 6786
  GetWriteCacheLine: 1563
  SlowRead: 1965009
  SlowWrite: 178801
  DisallowCacheRead: 61353
  DisallowCacheWrite: 120724
  InvalidateCache: 44822
 There's a nice reduction in 'InvalidateCache' events, but still quite some
 remain. Likely because memory mapper segment switches still triggers a lot of
 invalidations.

Metal Gear 2, intro demo
  NonCachedRead: 2172578
  NonCachedWrite: 2482304
  GetReadCacheLine: 2151049
  GetWriteCacheLine: 1844091
  SlowRead: 21551
  SlowWrite: 2482130
  DisallowCacheRead: 2814
  DisallowCacheWrite: 4924
  InvalidateCache: 2350158
 Almost no improvement compared to baseline. Likely there's not much slot
 switching going on: the BIOS ROM stays permanently selected in page 0, the
 game ROM in pages 1 and 2, and the system RAM in page 3. The many
 invalidations are caused by mega-ROM bank switches.