[5a/7] Check alignment of <mem base=.. size=..>
In the current implementation the value of the 'base' and 'size' attributes in
<mem> tags in a hardwareconfig.xml file, must be aligned to cache-line
boundaries (256 bytes). All of our configs already obey that alignment, but we
didn't check for it. This patch adds such a check.

Later patches in this series will require a more strict alignment. Therefor
this patch makes that alignment configurable per device (even though for now
all devices still keep it at 0x100).