Fix(?) sound problem on windows
In our SDL2 port windows did not produce sound anymore, when using the default
SDL windows sound driver (WASAPI). It did work again when forcing the legacy
drivers (directsound or winmm).

The problem is that the WASAPI driver prefers (or only supports?) the AUDIO_F32
sample format (32-bit floating point) while openMSX always assumed the
AUDIO_S16 format (16-bit signed integer). I believe SDL1 did not yet have
support for floating point samples (but I didn't re-check).

Another part of the problem is that we used the (SDL1 backwards compatibility)
function SDL_OpenAudio() to open the audio device. The documentation states
that the desired and the obtained format may not be the same. OpenMSX checked
this difference for the sample frequency, but not for the sample format.

The solution is to instead use the new SDL2 SDL_OpenAudioDevice() (and related)
functions. This one allows to specify which items are allowed to be different
between the desired and the obtained format. And if there's a mismatch between
the request and what the driver supports, SDL itself will perform the needed
conversions.