fix several '-Wnon-virtual-dtor' compiler warnings (gcc)
This flag is not (yet?) enabled by default. This patch tries to fix all
warnings triggered by enabling this flag. None of these changes fix actual
bugs, but they do make the code more robust against possible future mistakes.
Unfortunately 3 false-positives remain, and those occur in header files that
are #included quite a lot, thus the same warning gets repeated many times. I
believe this should be improved in gcc, and I'll try to file an enhancement
request for it.

Some more context about the warning itself:

In C++ when you have a 'Derived' class that inherits from a 'Base' class, and
if you delete an object with dynamic type 'Derived' via a pointer that has
static type 'pointer-to-Base', then the destructor of the 'Base' class must be
virtual. If not, your program has undefined behavior.

The rule of thumb we follow in openMSX is:
 - For base classes, make the destructor either:
    * public and virtual, or
    * protected and non-virtual
 - Mark leaf classes (non-base classes) as 'final' classes.

Gcc has two warnings that help to enforce the above rule(s):
  -Wdelete-non-virtual-dtor  and  -Wnon-virtual-dtor
The former is part of -Wall, the latter is not. (As far as I understand it) the
former catches actual undefined behavior in your program (when you actually do
an invalid delete). The latter warns about possible problems if you would
delete via a Base-pointer (e.g. after refactoring the code).

IMHO -Wnon-virtual-dtor is a useful warning, but because of the 3
false-positives that trigger in our code, I don't think we should (already)
enable it.