Fix(?) 'kbd_code_kana_host_key' setting
This setting allows to specify a different host key for the MSX code/kana key.
Though because of a bug changing this setting (sort of) acted like an
*additional* host key for code/kana instead of replacing the default host key
(which is right-alt).

See also
  https://www.msx.org/forum/msx-talk/openmsx/altgr-not-working-under-windows-and-a-question-about-alt-codes-in-console?page=0

I'm not sure the problem in the above post is fixed. But at least with an
AZERTY keyboard layout, I can now type the Yen-symbol on a turbor by using the
host key combination 'Alt-GR + -' (that is the '-' key (next to the '0' key on
a US-QWERTY keyboard).