Fixed remaining '-Wnon-virtual-dtor' warnings on gcc
This is a follow-up on fced193b5ff6.

At the time I thought the remaining warnings were false-positives in gcc. But
it turns out gcc was right. The class really does have an accessible
non-virtual destructor. The destructor *is* protected, but it also has a
*friend* class which in principle could access it.

Now both gcc and cland compile warning-free with '-Wnon-virtual-dtor'. (But
this class does not yet enable that warning flag).