Add work-in-progress build definitions for Meson
I've been trying the Meson build system as a candidate to replace
the custom build system we have for openMSX. So far it seems feasible,
but I don't want to draw conclusions until I've tried cross-compilation
as well.

I'm committing this now because it doesn't interfere with the existing
build system and other people might want to try it.