Let machine_info device <MSXRom device> return a dict
This is a backwards incompatible change, which is a long standing TODO. I
changed all scripts that we ship that make use of the extended device info that
is given for ROM devices.

Next to returning a dict, it also returns the (patched) sha1 of the device.
With that I could enhance the rom_info script to actually show information
about the patched ROM (if available) and show that the ROM has been
soft-patched by openMSX. This was the ultimate goal of this patch.

Already prepared the release notes to include a backwards-incompatibility
warning, so we won't forget.