Change output of 'machine_info device <dev>' to dict
Previous patch only returned a dict for the 2nd element (the extra device
info). After some discussion with Wouter, we decided that now that we break the
backwards compatibility anyway, we can just go all the way and return a full
dict for all info. For devices without extra info, this means that instead of
just the <type> you get a dict with key "type" and value "<type>".
This is often easier in scripts and at least more consistent. It also allows
easy expansion of the base device info, without breaking backwards
compatibility later (at least, if the users use the dict command to fetch the
info!).