Extended TclObject list/dict creation functions
Added a more convenient way to add multiple elements to a Tcl list at once.
This allows to replace:
   myObject.addListElement("foo");
   myObject.addListElement(calcAmountOfFoo());
with:
   myObject.addListElement("foo", calcAmountOfFoo());

Also added a way to directly construct a tcl dict (this can be more efficient
than creating a list of key/value pairs). So the above example can also be
written as:
   myObject.addDictKeyValue("foo", calcAmountOfFoo());

Or you can add multiple entries with a single call:
   myObject.addDictKeyValues("foo", calcAmountOfFoo(),
                             "bar", calcAmountOfBar());

This patch is mostly a cleanup patch. But possibly in script-heavy situations
it could also improve performance.