Allow to convert an Event to a Tcl list
This is a follow-up on e3b649ab20b47b. It fixes some TODO's from that commit.

Basically this avoids various string processing operations:
 - split string into list / join list into string
 - convert int from/to string
Now the information is created as a Tcl list and kept that way the whole time.

This also simplifies the interface of the Event class a bit (change an out-
parameter into a return value).