clang-tidy-8: fix 'bugprone-too-small-loop-variable' bugs
These are all theoretical bugs. A loop like
   for (unsigned i = 1; i < tokens.size(); ++i) {
with 'tokens' a std::vector, goes indeed wrong when there are more than 4
billion (2^32) items in the vector. This never happens in practice. Similarly
we will never have more then 4 billion pixels on a line or more than 4 billion
parittion on a harddisk.

Though fixing these bugs is easy and also the generated x86 code after the fix
is even a tiny bit faster. That is because the old (wrong) generated asm code
requires an extra explicit instruction to truncate a 64-bit value to 32-bit.