Ran include-what-you-use (iwyu) and remove some unnecessary #include's
This should result in (slightly) faster compile speeds, and more importantly
less incremental compilation when e.g. some header file is changed.

After this change, the code compiles fine for me, but there is a chance I broke
compilation on other platforms. If that's the case it should be easy to fix
again.