Use SDL_WINDOW_FULLSCREEN_DESKTOP instead of SDL_WINDOW_FULLSCREEN
When switching to fullscreen, instead of changing the monitor resolution (e.g.
to 640x480), this keeps the native desktop resolution (e.g. 1920x1080) and
scales+letterboxes the image.

For the SDL renderer this change is completely transparent (SDL fully handles
the extra scaling step internally). For the openGL renderer we have to do this
ourselves. When needed we render to a texture and then draw that texture to the
full desktop resolution.

This may be a very small step towards the scaling changes Maarten is proposing.
Though this is 100% a coincidence.