Use NPOT textures in the GLImage class
NPOT = Non-Power-Of-Two

Before this patch the GLImage class would round the x/y-dimensions of the image
up to the next power of two. Now we skip this.

This code was written around 2004 and at that time not all gfx cards had
support for NPOT texture. Nowadays it shouldn't be a problem. IIRC NPOT is
required in openGL-2.0 and also openGL-ES-2.0 (but with some restrictions).

This also fixes some graphical glitches that became (more) apparent after the
recent full-screen refactoring when setting
 - renderer = SDLGL-PP
 - scale_factor = 1
 - full_screen = true
The problem was that CLAMP_TO_EDGE on 'emulated' npot textures doesn't have a
real edge to clamp to on the bottom/right. 'Real' npot textures solve that.