Fix OSD menu in SDLGL-PP/fullscreen
More specifically the '-clip' property on OSD widgets didn't work properly
anymore.

Problem was that glScissor() works with clip-space coordinates, which comes
after the viewport-transformation. In other words: now that in fullscreen mode
we can have more (clip-space) pixels than we have logical pixels in the
viewport-space, we need to (manually) transform the parameters we pass to
glScissor().

Also made a small performance tweak in the OSDWidget::paint..() functions: now
that we have the optional<> helper class, we can use that to allocate the
{SDL,GL}ScopedClip object on the stack rather than heap-allocating it.