Fix screenshot with SDLGL-PP/fullscreen
In our openGL renderer, taking a screenshot makes use of the glReadPixels()
function to get access to the actually rendered pixels. But now in fullscreen
mode there are possibly more physical than logical pixels. But because
glReadPixels() only knowns about physical pixels and the old screenshot code
only knew about logical pixels, we ended up with a PNG that contained only part
of the MSX screen.

Fixed by making a more explicit distinction between physical and logical screen
size in the OutputSurface class. And using the physical size in the screenshot
code.

A consequence of this change is that now making a screenshot in windowed versus
fullscreen mode results in a PNG of different size.