Fix behaviour of "wrong" port config
As per TomH's CLK:
* a port configured as input will appear to be outputting 0xff to anyone
  elsewhere on the bus
This is now also implemented in openMSX. With this, the same behaviour as on a
real VG 8020/00 appears with the buggy XRacing 1.0 game.

The datasheet of the AY-3-8910 can be interpreted in such a way that it
describes this behaviour.

Fix implemented by Wouter, who asked me to commit it after testing with XRacing
1.0.