remove 'dead' code
    the condition:
    ```
            ((type == ROM_ASCII16) &&
                (typeGuess[ROM_GENERIC_8KB] == typeGuess[ROM_ASCII16]))
    ```
    is always `false` due to the fact that `typeGuess[ROM_GENERIC_8KB]`
    is always zero and `type` is initialized as `ROM_GENERIC_8KB` -
    would change to `ROM_ASCII16` only if `typeGuess[ROM_ASCII16]` > 0
    which makes second part of condition `false`

    PS i wanted to track down to commit which has changed some logic and created
    this leftover but it wasn't possible - some part of commit history was lost
    during move from svn to git and this code was already there back in 2004

--

[Wouter]: Small update.

I don't think we lost any history during the svn->git migration. But this
mapper detection code moved several times between files, so it's a bit hard to
follow. Anyway, I tracked it down to this commit:

    commit 5c31756ac7d5f8118f6b4ccb7ade3b4a54ace6eb
    Author: David Heremans <turbor@users.sourceforge.net>
    Date:   Mon Mar 25 07:35:16 2002 +0000

        More help in CommandLineParser
        Cleaned up and improved auto type detection in gamecartridge.

Before this patch the condition you found made sense, after it's indeed dead.

We could try to restore the original intention of this condition. But it's
anyway only a heuristic, and it's been working good enough like this for almost
17 year. So it's probably better to keep it as is, and apply your
simplification.

Thanks.