Initial emulation of Toshiba_HX-F101PE
This is an external Single Sided Double Density 3.5" disk drive. The
'connectionstyle' is different from existing styles. I reverse engineered it
from looking at the disk ROM code. In this MRC thread:
  https://www.msx.org/forum/msx-talk/hardware/toshiba-hx-f101-external-floppy-disk-drive-problem?page=4
'zeilemaker54' independently came to a similar conclusion.

The function of the register at location 0x7ff6 is still unclear to me. For the
moment it's emulated as a single bit storage without any further attached
functionality (see comments in code for more details). But so far it seems to
work fine.

This patch also fixes a crash when writing a sector beyond the end of the disk.
This for example happened when you use a 360kB disk image which has a
double-sized/80-track boot-sector and media-id (the current heuristic
intentionally detects such disks as double sided/40-tracks, see comments in
code for details). And then you try to format this disk.