Added MemoryBufferFile (+unittest)
This allows to create an openMSX 'File' object, but the content of that file is
not stored on the filesystem, instead it comes from some memory buffer.

I suspect that (for now?) this functionality will only be useful to write
unittests. And I intent to use it in a later patch to test the (recently
refactored) WavData class.