Major speedup in loading wav files
I still had some concerns about the new WavData implementation: it did one file
read() call per sample. Instead of one big read for the whole file. So I
measured it: reading a ~30MB wav file took ~3.5seconds! That's unacceptably
slow!

I changed the implementation to read the whole file at once (so 1 system call
instead of millions). Actually I don't read() the file but mmap() it. That
avoids the memory overhead for the temporary buffer. (Or better said, the RAM
pages that hold the content of the file are shared with the operating system's
disk cache).

The same test as above now only takes ~14ms (250x faster!). This is acceptable.

(And the unittest I added in the previous patch still passes).