Integrate WAV parser and filter loop
In WavData we loop over the WAV data to (possibly) convert 8->16 bit, or
(possibly) convert endianness. Later in WavImage we loop over this converted
data again to apply a DC filter. This data can be multiple MBs in size,
which is often bigger than the host CPUs L3 cache. This means that both loops
likely read/write from/to cold memory. It's possible to convert+filter in a
single pass (rather than 2 passes). This cuts the memory access overhead in
half.

Some measurements (on a ~30MB WAV file):
* before:
  - convert: 14ms
  - filter : 58ms
* after:
  - combined convert+filter: 64ms
That's only about 10% faster, which is less improvement than what I expected.
So I investigated further:
 - The filter used in WavImage is an IIR-type filter which means it's difficult
   to vectorize. And indeed gcc/clang emit about 20 x86 instructions for the
   convert+filter loop. And each iteration of the loop only processes 1 sample.
 - In the common case of a 16-bit mono wav, the input is relatively dense:
   there fit 32 samples in a single (64 byte) cache line.
 - Combined this gives several hundred x86 instructions per (input) cacheline.
   This seems enough to hide (most of) the memory latency.

So this patch only gives a modest speedup. But I decided to keep it anyway
because:
 - It doesn't really increase the complexity of the code.
 - In fact it allows to simplify the WavData interface.

Before the WavData class (had to) allow to modify the wave data after it had
been constructed. Now that's no longer needed and the full WavData interface
can become 'const'. This effectively makes WavData objects immutable.

In the future we might exploit this immutability to share WavData objects among
MSX machine instances (this e.g. (temporarily) happens when using the reverse
feature). But I have no immediate plans to work on this.