Fix issue#1106: Vertical adjust does not affect maximum horizontal interrupt line
I only verified that this patch makes the BASIC test program of the 1st post
in the bug report behave the same as my real turboR-GT. I didn't do much other
tests.
  https://github.com/openMSX/openMSX/issues/1106

The patch makes sense (to me). But I would appreciate it if someone more
familiar with this part of the VDP emulation (Maarten?) could review my patch.