Fix inaccuracy in DynamicClock::getFreq()
When using 'frequency=48000' and 'resampler=fast' and playing a PSG tune, I
could trigger the following assertion:
   openmsx: src/sound/ResampleLQ.cc:155: bool openmsx::ResampleLQDown<CHANNELS>::generateOutput(int*, unsigned int, param) [with unsigned int CHANNELS = 1; openmsx::EmuTime::param = const openmsx::EmuTime]: Assertion `(p + 1) < valid' failed.
With 'frequency=44100' (which I usually use) this doesn't happen, so I never
noticed this before.

For timekeeping in the audio code we use a DynamicClock object and set the
frequency of this clock to the host sample rate (48kHz in the above scenario).
The resolution of this clock is finite (it's based on EmuTime). A frequency of
48kHz cannot be set exactly. Instead the closest approximation is 47999.93Hz.
This is more than close enough. Though in some parts of the audio code
(specifically the ResamplerLQ code) we use the DynamicClock::getFreq() method,
and there 47999.93Hz got rounded down to 47999Hz instead of rounding to the
closest integer which would give back the original frequency.

This caused some parts of the calculation to be done with 47999Hz and others
with 48000Hz. This is still a very small difference. But when this small error
is accumulated over 1024 samples (a typical value for the 'samples' setting)
the error adds up to about 2%. This is still a small error which will be barely
audible (if at all). But it's enough to trigger an assertion (which checks for
buffer-index out-of-range).