Fix hangul font ROM on Daewoo CPC-400S
Apparently this machine has a 256kB font ROM and it is decoded with 7 bits for
the higher address, not 6 with an extra set of I/O ports like for JIS2.
See also https://www.msx.org/forum/msx-talk/openmsx/daewoo-cpc-400s-hanja-display-bug
The 256kB size is mentioned in the manual of this machine.

The extra bit is activated with the 'hangul' type for the Kanji device.