Use EPROM dumps with window for Daewoo CPC-400S.
Using EPROM dumps is more natural. The content is identical to what was before
in the XML file. The names of the 16kB areas are based on the names in the
manual of the machine.
Note that the original ROMs were strange: for instance, the subROM had an
unrelated 16kB block concatenated to it, that was actually the 2nd part of the
hangul ROM. This has now been normalized, using the names of the manual.