More refinements in MatraCompilation mapper.
After more serious testing, it appeared that some games were not working. They
were writing to the mapper registers. As the block offset register is always
written last by the software, I came to the conclusion that a value of 1 (and
probably also 0?) means that the mapper mechanism will be disabled. So I
changed the code to reflect this and now all games seem to work.

Implementation detail: I changed the standard value of the offset register back
to 2 (as was my first (not pushed) version). This way we don't need negative
values to check for a disabled mapper mechanism.