Tweak ROMHunterMk2
Manuel asked me to review this code. I think it was correct. (That is:
according to the reverse-engineered specification. I have no idea if that
specification is correct.) I only simplified the implementation a bit. I didn't
know how to test this, so hopefully this refactoring doesn't introduce any new
errors.