Re-apply serialization improvements.
This reverts an earlier revert:
    commit 2e1191cd8d6300a1d6a560ee74c413af28ae0e30
    "Revert some problematic patches for vs2015"

Or in other words it re-applies the commits:
    633507c2922f222ff6c4aeeca4e9ccc334a06302
    0a00e8a1256e7ce3ea85ee611c3762d3de159ca6
    0050c9688f2d73a338bfb2849fdedc226b9aa669
    ddf6bda1112bb662f8293baab3a8ee4a1adffe86
    345abade7a7231b2ee249e62b541adb18989b9b8
    25bf9193025aacef763e849b72e32aae1413b70f
(+ make adjustments for code changes that happened in the mean time).

See the above commits for lots more details. Very short summary:

Replace stuff like
    ar.serialize("foo", foo);
    ar.serialize("bar", bar);
with
    ar.serialize("foo", foo,
                 "bar", bar);

This changes nothing for XML based savestates, but makes in-memory savestates
(used by the reverse system) more efficient (because serialize() has a more
global view on what needs to be done). The reason why it was reverted is that
it could crash older versions of visual studio. But that was 4 years ago. Let's
try again now.