Use c++17 static_assert without a message
The 2nd parameter (the error message) in a 'static_assert' is optional in
c++17. This is useful because often the 1st parameter (the expression) is
expressive enough. For example:

c++14:
    static_assert(sizeof(Foo) == 16, "Foo must have size 16");
typical workaround (c++14)
    static_assert(sizeof(Foo) == 16, "");
c++17:
    static_assert(sizeof(Foo) == 16);

This is intentionally a very simple first c++17 patch, because I'd like to
verify that all (automatic) build loops are correctly configured for c++17.