Use c++17 constexpr-if to simplify tag-dispatch helper stuff
C++17 has a new feature called constexpr-if:
    if constexpr (<expr>) {
        <then-stuff>
    } else {
        <else-stuff>
    }
This is much like a regular if, but with these differences:
- <expr> must be a constant-expression, that is the value must be known at
  compile-time.
- Depending on the value of <expr> either the then or the else branch is kept,
  the other branch is _completely_discarded_ (apart from some very basic syntax
  checking).

This may still seem the same as a regular-if which you pass a constant
expression: that one will also discard the non-taken branch, but only _after_
the optimizer has run. Thus for a regular-if the discarded branch must still
contain valid c++ code. E.g. all templates in that branch are still being
instantiated. Sometimes those instantiations are problematic, and preventing
those requires ugly workarounds. See this patch for some concrete examples.