Use 'if constexpr (..)' where possible
The statements
    'if (<expr>) { <body> }
and
    'if constexpr (<expr>) { <body> }
are largely the same, but with these two differences:
* <expr> must be a constant-expression (possible to evaluate during
  compile-time).
* In case the then- or else-block is not selected, there are less strict
  checks on the <body> of that block.

Here's an example:
    template<typename Pixel>
    void generic_version(Pixel* line)
        #ifdef __SSE2__
        if (sizeof(Pixel) == 4) {        // better: if constexpr
            optimized_SSE_version(line);
            return
        }
        #endif
        ...
    }
This (templated) function can work with various Pixel types (e.g. 32bpp
and 16bpp). We have an optimized SSE version, but only for 32bpp. When
this template gets instantiated with a 16bpp Pixel type, the condition
in the 'if' evaluates to false, so the body won't execute at run-time.
Nevertheless the body must still be compilable. This means we need an
overload of optimized_SSE_version() which accepts the 16bpp type, even
though this function will never be called. If we replace 'if' with 'if
constexpr' this dummy overload is no longer required.

In practice 'if constexpr' is mostly useful in template functions. But
we also use it in combination with 'global constants'. For example:
    if constexpr (OPENMSX_BIGENDIAN) { ...

Apart from the less-strict checks in 'if constexpr' there shouldn't be
much difference. For 'if constexpr' the compiler is _required_ to not
include the non-selected branch in the code. For 'regular if' it depends
on the compilers optimizer, but in practice all optimizers do remove
dead blocks (though this happens later in the compiler pipeline). In
theory this should lead to identical generated code, in practice I did
see minor differences.

In theory using 'if constexpr' could result in faster compile-times
(because the compiler can/must remove code sooner). However in our
current code base this effect is not measurable.

Also (IMHO) using 'if constexpr' expresses the intend of the code
better: in some cases we really want to make decision at compile-time
rather than run-time.