Use 'standard' names for some of our Math utility functions
C++20 will have the functions 'std::ispow2()' and 'std::ceil2()'. These
respectively correspond to our functions 'Math::isPowerOfTwo()' and
'Math::powerOfTwo()'.

This patch renames our functions to match the standard functions (but still in
the Math namespace). In the future we can then switch to the 'std' version and
drop our implementation.

The latest gcc/clang versions already implement these std functions. I verified
that their implementation is more or equally efficient than our version. That's
nice for if we later switch to those. But for now I didn't bother tweaking our
versions because they are not used in performance critical code.