Suppress Coverity#1484550 uninitialized scalar field
Coverity complained that the constructor did not initialize all member
variables. This was intentional: only when some boolean member variable was set
to true, the value of some other members mattered.

Though this concept of optionally initializing something plus having a boolean
that indicates whether it is initialized is more clearly expressed by using
'std::optional' (or our pre-c++17 implementation of it).