Annotate value-getter methods on Setting subclasses as non-throwing
Our Setting classes (e.g. IntegerSetting, BooleanSetting, ...) are based on Tcl
variables. In general converting the value of a Tcl variable to an integer,
boolean, ... may result in an error (e.g. trying to interpret the value
"foobar" as an integer is an error). Though in the context of our Setting
(sub)classes we enforce that the corresponding Tcl variable can only take on
valid values for that setting type.

Applied to the c++ code:
Let's take the BooleanSetting::getBoolean() method as an example. The
implementation is:
    return getValue().getBoolean(getInterpreter());
The first part 'getValue()' returns the TclObject that holds the value of the
setting. The second part 'getBoolean(getInterpreter())' tries to convert that
value to a boolean (according to the Tcl rules). When the value has no valid
boolean interpretation it throws an exception (and it needs the Interpreter
object to construct a good error message). Though in the larger context we know
this value will always have a valid boolean interpretation.

So these value-getter methods are logically non-throwing, but their
implementation might call potentially throwing functions. These are good
functions to explicitly annotate with 'noexcept'. This (c++11) keyword tells
the compiler that the annotated function will never throw an exception. If an
exception does escape from such an annotated function then std::terminate()
gets called. In our case that's good because it means there's a bug in the
setting value handling code.

In some cases these annotations allow the compiler to generate better code
(e.g. because there are fewer potential control-flow paths). But the main
motivation for this patch is to help static analyzers like Coverity. Before
this patch e.g. Coverity#1463930 complained about a possibly throwing
destructor, but with extra knowledge about the behavior of the Settings, those
exceptions cannot happen. Hopefully this patch fixes those false-positives.