Made XMLElement::findChildWithAttribute() more robust
This method searches for a (direct) child node which:
- has a specific name
- and an attribute with specific name and value

(Before this patch) this method would throw an exception when:
- there was a child with the correct name
- but it did not have an attribute with the specified name

On the other hand when:
- no child with the requested name is found
- or there is a child with the request name and attribute name but it has the
  wrong attribute value
then this method simply returns "not found" instead of throwing an exception.

This throwing behavior was never the intention, so this patch fixes that. None
of the current contexts where this method is called could trigger this throwing
behavior. But it's anyway cleaner to fix it.

This may also fix some false-positive Coverity reports about possible uncaught
exceptions.