Pass inputRate parameter to SoundDevice constructor
As a rule (e.g. to avoid UMR) each SoundDevice subclass should call
SoundDevice::setInputRate() in its constructor. The only thing that call does
is set the 'inputSampleRate' variable in the base class.

All current code already followed this rule, but still it's better to enforce
such rules in the code structure. In this case that's possible by making
'inputRate' a constructor parameter of SoundDevice.

This also fixes Coverity#1484539 which complained that SoundDevice didn't
initialize its 'inputSampleRate' member variable.