Don't use (stateful) iostream formatting in CPUCcore
Coverity#1191180 warns that the IO formatting in CPUCore::cpuTracePost_slow()
does not restore the formatting state. This could lead to a situation where,
depending on whether cpu tracing is enabled or not, a totally unrelated print
statement gets differently formatted. (IMHO having stateful formatting in the
std iostreams was a design mistake).

It is possible to save/alter/restore the formatting state, but instead this
patch switches from std iostream formatting to our strCat utility to do the
formatting. (At the time we originally wrote the cpu tracing code we did not
have strCat yet). (In the future a possible alternative could be c++20
std::format).