Modernize use of fdopen()
The "with" block is a simpler way to ensure the file always gets
closed.

I removed the resetting of "buf" at the end, since I don't see any
good reason for it: it's not likely for it to get accidentally reused
and it's not a problem if it doesn't immediately become available for
garbage collection.