Added -once flag to 'debug set_bp' (and related) commands
This feature was requested by 'grauw'.

Sometimes you want to set a breakpoint (or watchpoint, ...) that should only
trigger once. In gdb-terminology this is called a temporary breakpoint.

In principle this was already possible by removing the breakpoint in the
breakpoint's callback command. Though this is cumbersome because the
breakpoint-id that's required to remove the breakpoint is only known after the
breakpoint is created (and thus also after the callback command is already
set). A possible solution for single breakpoints is to store the breakpoint-id
in a global Tcl variable. But if you want to manage a collection of breakpoints
this becomes messy.

This patch adds an option '-once' flag to the commands:
  debug set_bp
  debug set_watchpoint
  debug set_condition
  debug probe set_bp
When a breakpoint (watchpoint, ...) with this flag triggers, it's automatically
removed. In other words: it only triggers once.