Using make_xxx() factory functions is often no longer needed in c++17
Suppose we want to construct a pair from an int and a double. Pre-c++17 this
could be done like:
   int i = ...;
   double d = ...;
   std::pair<int, double> p(i, d);

Though this requires repeating the types 'int' and 'double'. Pre-c++17 already
offered a helper function std::make_pair to avoid this repetition:
   auto p = std::make_pair(i, d);

std::make_pair is a templatized function. In the example above we actually
invoke std::make_pair<int, double>, but because of function template argument
deduction it's not required to explicitly specify 'int, double'.

C++17 adds a feature called 'Class Template Argument Deduction' (CTAD). This
allows to deduce the template parameters of a class from the constructor
arguments. The example above then becomes:
   std::pair p(i, d);