[2/2] use constexpr to define constants
Before this patch we mostly used 'static const' to define constants. This patch
replaces 'static const' with 'constexpr'.

For example:
    static const int MY_CONSTANT = 123;
->
    constexpr int MY_CONSTANT = 123;

This is a little shorter, but that's not the real reason for this change.

C++ 'constant' variables can have these properties:
 a) Once initialized, their value cannot change.
 b) The value is already known during compile-time (rather than the result of a
    run-time calculation).

In C++ when you write:
    static const <type> MY_CONSTANT = <some-expression>;
it can mean either 'a' or 'a+b', depending on <type> and <some-expression>. On
the other hand if you write:
    constexpr <type> MY_CONSTANT = <some-expression>;
it must mean 'a+b', 'a'-only would be a compile error. (Side-note: there are
proposals for future c++ standards to initialize variables with property
'b'-only).

In the openMSX code we practically always want 'a+b', so let's use the syntax
that expresses those semantics.

To make a connection with the previous patch: a 'constexpr' variable implies
'inline'. So this patch also allowed to cleanup some (now redundant) constant
definitions from .cc files (e.g. see the removed comments in
src/fdc/RawTrack.cc)